+++ package.json
@@ -35,7 +35,8 @@
     "test:types": "tsc -p . --noEmit",
     "selenium": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/selenium.ts",
     "update-bcd": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/update-bcd.ts",
-    "find-missing-features": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/find-missing-features.ts",
+    "feature-coverage": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/feature-coverage.ts",
+    "find-missing-features": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/feature-coverage.ts",
     "find-missing-reports": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/find-missing-reports.ts",
     "add-new-bcd": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/add-new-bcd.ts",
     "release": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/release.ts",
