+++ scripts/feature-coverage.test.ts
@@ -1,6 +1,6 @@
 //
-// mdn-bcd-collector: unittest/scripts/find-missing-features.test.ts
-// Unittest for the missing features finder script
+// mdn-bcd-collector: unittest/scripts/feature-coverage.test.ts
+// Unittest for the feature coverage report script
 //
 // © Gooborg Studios, Google LLC
 // See the LICENSE file for copyright details
@@ -12,7 +12,7 @@ import fs from "fs-extra";
 
 import bcd from "../unittest/bcd.test.js";
 
-import {traverseFeatures, getMissing} from "./find-missing-features.js";
+import {traverseFeatures, getMissing} from "./feature-coverage.js";
 
 const tests = await fs.readJson(
   new URL("../unittest/tests.test.json", import.meta.url),
