+++ scripts/report-stats.ts
@@ -20,7 +20,7 @@ import {Report, ReportStats} from "../types/types.js";
 import {BCD_DIR} from "../lib/constants.js";
 import {parseUA} from "../lib/ua-parser.js";
 
-import {findMissing} from "./find-missing-features.js";
+import {findMissing} from "./feature-coverage.js";
 
 const {default: bcd}: {default: CompatData} = await import(
   `${BCD_DIR}/index.js`
